python/companies/requirements.txt
grpcio==1.10.0
pika==0.11.2
protobuf==3.5.2
sqlalchemy==1.2.14
psycopg2==2.7.4